requirements.txt
google-generativeai>=0.2.2
flask>=2.0.0
